Move `test_store.py` to `test_stores/test_sql.py`
